third_party/blink/web_tests/platform/generic/http/tests/websocket/handshake-challenge-randomness-expected.txt

Handshake request should contain random challenge values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS challenge1 === challenge2 is false
PASS successfullyParsed is true

TEST COMPLETE
